+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2021-1-23
+## [Unreleased] - 2021-1-25
 
 ### Added
 * A factory method, ConstantRestartSchedule.createRestartSchedules, for creating multiple restart schedules.
-* A factory methods, VariableAnnealingLength.createRestartSchedules, for creating multiple restart schedules.
+* Factory methods, VariableAnnealingLength.createRestartSchedules, for creating multiple restart schedules.
+* Factory methods, LubyRestarts.createRestartSchedules, for creating multiple restart schedules.
 
 ### Changed
 * Refactored TimedParallelMultistarter and TimedParallelReoptimizableMultistarter to remove redundancy.